Add files according to midproject
We have covered main parts like registration/loging in, posts, editing, friends, notifications and all other stuff. Later will upload commits one by one with information of whoms work is in which part. We have been waiting for this moment very long time, so that we have forget about commiting the work on git. Sorry for that.